AuthorSelector: Move SwitcherShell to dedicated file
